requirements.txt
numpy==1.24.3
streamlit==1.30.0
redlines==0.4.2
ipython==8.15.0
streamlit-option-menu==0.3.6
bs4==0.0.1
google-generativeai==0.3.2
keras==2.14.0
tensorflow==2.14.0
Pillow==9.4.0
